+++ src/components/sections/UserManagement.stories.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import UserManagement from './UserManagement.jsx';
+
+export default {
+  title: 'Corporate/Sections/UserManagement',
+  component: UserManagement
+};
+
+export const Default = {};